prev no test

+++ task_1/funcs.hpp
@@ -1,8 +1,8 @@
 #ifndef TRIGFUNCS_H
 #define TRIGFUNCS_H
 
-double sin(double, int);
-double sinh(double, int);
+double sin(double, double);
+double sinh(double, double);
 double square_rootPD(double, double);
 double pow(double, int);
